- Redis command - Map command - Change underlying type for IntegerParameter to Long - More reasonable dict command - Make with upvar 1 everything but target

+++ src/main/java/com/basicer/parchment/parameters/IntegerParameter.java
@@ -14,7 +14,7 @@ public class IntegerParameter extends Parameter {
 	}
 
 	@Override
-	public Class<Integer> getUnderlyingType() { return Integer.class; }
+	public Class<Long> getUnderlyingType() { return Long.class; }
 	
 	
 	public Double asDouble(Context ctx) {
@@ -48,4 +48,6 @@ public class IntegerParameter extends Parameter {
 	public int intValue() {
 		return self.intValue();
 	}
+
+
 }
